Update pdfsearch extension
- Update CHANGELOG.md
- Refactor failure handling in search
- Add file validation for search component
- Update comments for swift files
- Delete .vscode
- Edit .gitignore
- Fix opening to wrong page number
