+++ extensions/pdfsearch/.gitignore
@@ -13,3 +13,4 @@ compiled_raycast_swift
 .DS_Store
 
 .build/
+.vscode/
