+++ extensions/pdfsearch/CHANGELOG.md
@@ -1,5 +1,10 @@
 # PDFSearch Changelog
 
+## [1.0.2] - 2024-02-01
+
+- Fix PDF opening at page number corresponding to page index.
+- Only load files in directory when collection is selected.
+
 ## [1.0.1] - 2024-01-22
 
 - Resolved issue where the deleting collections throws an error.
