+++ extensions/pdfsearch/src/views/search.tsx
@@ -6,6 +6,7 @@ import { chmod } from "fs/promises";
 import { ExecaChildProcess, execa } from "execa";
 import path from "path";
 import { Collection, Document } from "../type";
+import { getValidFiles } from "../util";
 
 export default function SearchCollection(props: { collectionName: string }) {
   if (!props.collectionName) {
@@ -21,15 +22,23 @@ export default function SearchCollection(props: { collectionName: string }) {
     const index = (await LocalStorage.getItem(props.collectionName)) as string | undefined;
     if (!index) {
       showFailureToast(`Couldn't find collection ${props.collectionName}!`);
-      throw new Error(`Failed to get collection ${props.collectionName}!`);
+      return;
     }
+
     const collection = JSON.parse(index) as Collection;
+    const validFiles = getValidFiles(collection.files);
+    if (validFiles.length === 0) {
+      showFailureToast("No supported files found!");
+      return;
+    }
+    collection.files = validFiles;
+
     showToast({
       style: Toast.Style.Success,
       title: "Loaded",
       message: `Loaded collection ${props.collectionName}`,
     });
-    return collection || {};
+    return collection;
   });
 
   const searchFiles = useCallback(
@@ -67,7 +76,7 @@ export default function SearchCollection(props: { collectionName: string }) {
     delay 1
     tell application "System Events"
         keystroke "g" using {option down, command down}
-        keystroke "${page}"
+        keystroke "${page + 1}"
         keystroke return
     end tell
     `;
